requirements.txt
httpx==0.27.2
asyncio-throttle==1.0.2
opencc-python-reimplemented==0.1.7
